+++ static/locales/en.json
@@ -57,8 +57,7 @@
 		"title": "Collabs",
 		"create": "Create collab",
 		"update": "Update collab",
-		"export_csv": "CSV Export",
-		"export_assets": "Asset Export",
+		"export": "Export",
 		"ongoing": "Ongoing",
 		"no_ongoing": "There are no ongoing collabs",
 		"upload_logo": "Upload logo",
